+++ lib/forge_sdk/protocol/display/multisig.ex
@@ -3,12 +3,13 @@ defimpl ForgeSdk.Display, for: ForgeAbi.Multisig do
   Implementation of `Display` protocol for `KVPair`
   """
 
-  def display(%{signer: addr, signature: sig, data: data}, _expand? \\ false) do
+  def display(%{signer: addr, pk: pk, signature: sig, data: data}, _expand? \\ false) do
     addr = if String.valid?(addr), do: addr, else: Base.url_encode64(addr, padding: false)
     sig = Base.url_encode64(sig, padding: false)
 
     %{
       signer: addr,
+      pk: Base.url_encode64(pk, padding: false),
       signature: sig,
       data: data
     }